Load web font in head to prevent flash of incorrectly styled content
When loading the Typekit font at the bottom of <body/>, the font appearance
changes noticeably when the page is first loaded and the Typekit font
has not been cached previously.
Loading the font at the top increases the time until page content is
shown (when cache empty), but reduces irritation.