build: revert model_mommy removal
This package was removed due to the addition of a new similar package,
but there are still test dependencies on it. For now, this commit adds
the package back.

Note: this package seems to be deprecated due to a naming issue, we may
consider switching to model-bakery. [link](https://pypi.org/project/model-mommy/)